feat : roomList Service 기능 구현 룸 목록을 찾을 수 없는 ErrorCode 생성

+++ src/main/java/com/example/Waffle/repository/RoomRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface RoomRepository extends JpaRepository<RoomEntity, Long> {
 
-    List<RoomEntity> findAllByGroupId(int groupId);
+    List<RoomEntity> findAllByGroupId(Long groupId);
 }
